fix: Fix all warnings and errors in codebase
## Fixed Warnings:

### Unused Imports (3 fixed):
- Remove unused  import from src/networking/plugins/mod.rs
- Remove unused  import from src/security/network_policy.rs
- Remove unused  import from src/security/security_context.rs

### Deprecated Type Alias (1 fixed):
- Add  to  in migration/mod.rs
  to suppress warning about MaestroK8sClient

### Unused Fields (3 fixed):
- Remove unused  field from ResourceQuotaBuilder in src/security/resource_quota.rs
- Add  with TODO comments to unused fields in KubeJobStep
- Add  with TODO comments to unused fields in KubePodStep

### Unused Test Code (1 fixed):
- Wrap MockStorage struct with  so it's only compiled during tests

## Clippy Warnings (7 fixed):
- Add  for forward compatibility
- Remove unnecessary  cast in service.rs
- Replace unnecessary unwrap with map() pattern
- Add  derive to TLSConfig, ExecutionMode, and CheckpointFrequency
- Add  for PodSecurityContextBuilder and ContainerSecurityContextBuilder
- Replace  with
- Add  where required
- Add  for drop order

All TODO comments clearly indicate that unused fields in KubeJobStep and KubePodStep
are pending implementation according to specifications.

+++ src/networking/service.rs
@@ -142,7 +142,7 @@ impl ServiceBuilder {
             .into_iter()
             .map(|p| K8sServicePort {
                 name: p.name,
-                port: p.port as i32,
+                port: p.port,
                 target_port: Some(
                     k8s_openapi::apimachinery::pkg::util::intstr::IntOrString::Int(p.target_port),
                 ),
